Remove superfluous include sys/time.h lines and other cleanup
