+++ ext/rbkit_sampling_profiler.h
@@ -1,6 +1,5 @@
 #ifndef RBKIT_SAMPLING_PROFILER
 #define RBKIT_SAMPLING_PROFILER
-#include <sys/time.h>
 #include <stddef.h>
 
 typedef struct _rbkit_frame_data {
